backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
google-generativeai
python-dotenv
tensorflow
pandas
scikit-learn
joblib
numpy
psycopg2-binary
lightgbm
